app/requirements.txt
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11.14
django-cors-headers==2.4.0
django-extra-fields==1.0.0
django-filter==2.0.0
django-rest-framework-base64-fields==0.1.3
djangorestframework==3.8.2
dnspython==1.15.0
enum34==1.1.6
eventlet==0.24.1
greenlet==0.4.14
gunicorn==19.9.0
idna==2.7
Markdown==2.6.11
monotonic==1.5
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
whitenoise==3.3.1